update consume json data

+++ lib/models/Restoran.dart
@@ -48,7 +48,7 @@ class Restaurants {
     description = json['description'];
     pictureId = json['pictureId'];
     city = json['city'];
-    rating = json['rating'];
+    rating = json['rating'].toDouble();
     menus = json['menus'] != null ? Menus.fromJson(json['menus']) : null;
   }
   String? id;
